Fix SMS memory leak and safe mobile data termination

+++ Src/backend/utils/network_manager.cpp
@@ -46,7 +46,9 @@ void NetworkManager::turnOnMobileData() {
 }
 
 void NetworkManager::turnOffMobileData() const {
-    pppdProcess->terminate();
+    if (pppdProcess) {
+        pppdProcess->terminate();
+    }
 }
 
 void NetworkManager::onNewPPPConnection() {
